♻️ Refactor: Artist components
