+++ client/src/features/artist/components/Artist.tsx
@@ -10,7 +10,7 @@ import styles from '../../../ui/PlayHeader/PlayHeader.module.scss';
 import PlayButton from '../../../ui/PlayHeader/PlayButton.tsx';
 import TransparentButton from '../../../ui/Button/TransparentButton.tsx';
 import { followUser, unfollowUser } from '../../auth/userThunks.ts';
-import UploadSongDialog from './uploadSong/UploadSongDialog.tsx';
+import UploadSongDialog from './forms/uploadSong/UploadSongDialog.tsx';
 import NotFound from '../../../ui/StatusScreens/NotFound.tsx';
 import ServerError from '../../../ui/StatusScreens/ServerError.tsx';
